Add basic Widget and Signal builder APIs
For Widget classes, you have to specify how to draw the widget,
Widget classes also store bunch of geometry information, for later
Layouts implementation.

For Signal classes, it's just for binding Widget-specific methods onto
a public Signal, later we can emit the signal to all the bind methods
(with optional arguments) and execute the methods.

The Rectangle classes are not finished, just committing this, so I
could get it on my laptop lol